Move the test to ParquetPartitionDiscoverySuite
