+++ sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/FileIndexSuite.scala
@@ -31,7 +31,7 @@ import org.mockito.Mockito.{mock, when}
 
 import org.apache.spark.SparkException
 import org.apache.spark.metrics.source.HiveCatalogMetrics
-import org.apache.spark.sql.{QueryTest, Row, SparkSession}
+import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.util._
 import org.apache.spark.sql.functions.col
 import org.apache.spark.sql.internal.{SQLConf, StaticSQLConf}
@@ -39,7 +39,7 @@ import org.apache.spark.sql.test.SharedSparkSession
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
 import org.apache.spark.util.KnownSizeEstimation
 
-class FileIndexSuite extends QueryTest with SharedSparkSession {
+class FileIndexSuite extends SharedSparkSession {
 
   private class TestInMemoryFileIndex(
       spark: SparkSession,
@@ -518,21 +518,6 @@ class FileIndexSuite extends QueryTest with SharedSparkSession {
       }
     } finally {
       SQLConf.get.setConf(StaticSQLConf.METADATA_CACHE_TTL_SECONDS, previousValue)
-    }
-  }
-
-  test("SPARK-34314: preserve partition values of the string type") {
-    import testImplicits._
-    withTempPath { file =>
-      val path = file.getCanonicalPath
-      val df = Seq((0, "AA"), (1, "-0")).toDF("id", "part")
-      df.write
-        .partitionBy("part")
-        .format("parquet")
-        .save(path)
-      val readback = spark.read.parquet(path)
-      assert(readback.schema("part").dataType === StringType)
-      checkAnswer(readback, Row(0, "AA") :: Row(1, "-0") :: Nil)
     }
   }
 }
